tính năng quản lý sân

+++ src/main/java/com/example/football_field_management/repository/AccountRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
     Optional<Account> findByEmail(String email);
 //    boolean existsByEmail(String email);
 //    boolean existsByPhoneNumber(String phoneNumber);
